requirements.txt
# Recruiter Intelligence System - Dependencies
# Version: 1.1

# Core
feedparser>=6.0.0
requests>=2.31.0
aiohttp>=3.9.0

# Content processing
beautifulsoup4>=4.12.0
lxml>=5.0.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0

# LLM APIs
anthropic>=0.40.0
openai>=1.0.0
google-genai>=1.0.0

# Utilities
python-dateutil>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
structlog>=24.0.0
tenacity>=8.0.0
pyyaml>=6.0.0

# CLI & Output
typer>=0.9.0
rich>=13.0.0

# API (Phase 5)
fastapi>=0.109.0
uvicorn>=0.27.0

# Enhanced data sources
edgartools>=0.20.0  # SEC EDGAR Form D API
gdeltPyR>=1.0.0     # GDELT news API

# NER & Text Processing
spacy>=3.7.0        # Fast NER extraction

# Deduplication & Matching
rapidfuzz>=3.0.0    # Fast fuzzy string matching

# Testing
pytest>=7.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.0.0

# Note: After installing, run:
# python -m spacy download en_core_web_lg
